fix(types): mypy from 24 errors to 0, and one typed row that removed thirteen of them
The CI lint job was red and I had been running only `ruff` all session — the repo
has `./scripts/lint`, which is ruff + mypy + pyright + the architecture
invariants, and CI runs THAT. Running the gate the repo documents would have
shown this at the first commit.

Baseline first, so the numbers mean something: 22 mypy errors before today's
work, 24 after. Two were mine. All 24 are now gone.

Thirteen were one missing declaration. `SymbolTable.read_for` and `.find`
returned `dict[str, object]`, so every caller wrote `int(entry["line"])` and mypy
rejected each one — `object` is not an `int`. They now return `SymbolRow` /
`FoundSymbol` (storage/rows.py) and the casts are unnecessary.

The rest, each a real statement about the code rather than a silencer:

- `Wire` (the embeddings send protocol) declared settable attributes, and its
  ONLY implementation is a frozen dataclass whose `endpoint` is a property — so
  nothing ever satisfied it. Now read-only, which is also all `_send` does.
- `shortest_path` said `list[dict[str, str]]` while every caller treated the
  result as `Hop`; that mismatch is what forced a `type: ignore` at the call
  site, now removed.
- `_direction` read `hop["code"]` through `or {}`, which widened it to `object`
  and lost `.get`. It checks for None instead.
- `jumpable`'s isinstance guard became unreachable once the row was typed.

Two invariants I broke while fixing it, both caught by the suite:

- `_symbols.py` went past the 100-line budget, so the row types moved to
  `storage/rows.py` — which also lets a CONTRACT name a row without importing
  storage.
- `contracts/node.py` importing that module pulled `storage/__init__` and with it
  numpy, breaking "importing megabrain does not load numpy". It is under
  TYPE_CHECKING now; the annotation is a string anyway.

Green: ruff, mypy, and the full suite on 3.10, 3.12 and 3.13.

Still red: pyright, 75 errors. It was 107 before today (this work removed 32), so
it is a standing v3 debt rather than a regression — v3 never ran CI, because CI
only triggers on master and v3 had no PR. Left for the 1.0.0 review to scope.

+++ src/megabrain/knowledge/graph/semantic.py
@@ -49,7 +49,7 @@ def _cosines(vectors: Matrix, rows: dict[str, int], order: list[str]) -> Matrix:
     kept = vectors[[rows[path] for path in order]]
     norms = np.linalg.norm(kept, axis=1, keepdims=True)  # pyright: ignore[reportUnknownMemberType]
     unit = kept / np.where(norms == 0, 1, norms)
-    sims = unit @ unit.T
+    sims = (unit @ unit.T).astype(np.float32)
     # A file is its own nearest neighbour at 1.0, which would fill every top-k
     # slot with self-edges and leave the lane empty.
     np.fill_diagonal(sims, -1.0)
